License the software under BSD-2-Clause AND 0BSD
For the most part this is a tautology as the 0BSD license is
(*almost) a subset of BSD-2-Clause. As the more the restrictive
license, the software is essentially made available under the
terms of BSD-2-Clause

However each individual file "may" carry it's only license header.
This allows for files to be made available under the terms of 0BSD alone.
Check the license header of the relevant file for specific terms.